% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean_names_swe}
\alias{clean_names_swe}
\title{Create standard compliant variable names.}
\usage{
clean_names_swe(data = df, factor_to_string = TRUE, sep_cols = FALSE)
}
\arguments{
\item{data}{A dataframe with non-standard names.}

\item{factor_to_string}{Convert factors to string.}

\item{sep_cols}{Seperate colums that contain both code and name into two variables.
Only for downloaded SCB-data.}
}
\value{
A dataframe \code{df} with standard names.
}
\description{
Create standard compliant variable names.
}
\examples{
df <- data.frame(ärVar = c(1:3), örÅr = c(4:6), region = c("0 Riket", "12 Skåne", "1280 Lund"))
clean_names_swe(df)
}
